requirements.txt
# app packages
boto3==1.24.71
botocore==1.27.71
certifi==2022.6.15.1
charset-normalizer==2.0.12
httplib2==0.20.4
idna==3.3
jmespath==1.0.1
loguru==0.6.0
numpy==1.23.3
oauth2==1.9.0.post1
oauthlib==3.2.1
pandas==1.4.4
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12